MERO-2006: S3 - Multiple instance support for s3server via systemctl.
                Support for s3server startup/shutdown via systemctl and also support for multiple
                instances.

+++ stops3.sh
@@ -1,15 +1,23 @@
 #!/bin/sh
-
-for i in `ls /var/run/s3server-*.pid 2> /dev/null`
+for s3unit in `systemctl list-units | grep s3server@  | grep "loaded active running"  | awk '{print $1}'`
 do
-PIDSTR=`cat $i`
-read -p "Shutdown s3 server daemon with pid $PIDSTR (y/n)? " answer
-case ${answer:0:1} in
-  y|Y )
+  systemctl stop $s3unit
+done
+
+out=`systemctl list-units | grep s3server | grep "loaded active running"  | awk '{print $1}'`
+if [ "$out" != "" ]; then
+  systemctl stop s3server
+fi
+
+PIDSTR=`cat /var/run/s3server.pid 2> /dev/null`
+if [ "$PIDSTR" != "" ]; then
+  read -p "Shutdown s3 server daemon with pid $PIDSTR (y/n)? " answer
+  case ${answer:0:1} in
+    y|Y )
       kill -s TERM $PIDSTR
       if [ "$?" != "0" ]; then
-        rm $i
+        rm /var/run/s3server.pid
       fi
   ;;
 esac
-done
+fi
